moved plugin and components into different directory
