+++ plugin/components/FigureHover.js
