+++ plugin/components/Pill.js
