+++ public/index.html
@@ -12,7 +12,6 @@
       <strong>We're sorry but <%= htmlWebpackPlugin.options.title %> doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
     </noscript>
     <div id="app"></div>
-    <pui-accordion></pui-accordion>
     <!-- built files will be auto injected -->
   </body>
 </html>
