--- src/components/Accordion.js
@@ -1,6 +1,0 @@
-import { createApp } from 'vue';
-import Accordion from './Accordion.vue';
-
-const app = createApp(Accordion);
-
-app.mount('#app');
