+++ src/main.js
@@ -1,6 +1,6 @@
 import { createApp } from 'vue';
 import App from './App.vue';
-import PuiPlugin from './plugin.js';
+import PuiPlugin from '../plugin/plugin.js';
 
 const app = createApp(App);
